requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
addict=2.4.0=pypi_0
anyio=3.3.4=pypi_0
argcomplete=1.12.3=pypi_0
argon2-cffi=21.1.0=pypi_0
attrs=21.2.0=pypi_0
babel=2.9.1=pypi_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
ca-certificates=2021.7.5=h06a4308_1
certifi=2021.5.30=py37h06a4308_0
cffi=1.14.6=pypi_0
charset-normalizer=2.0.7=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=pypi_0
debugpy=1.4.3=pypi_0
decorator=5.1.0=pypi_0
defusedxml=0.7.1=pypi_0
deprecation=2.1.0=pypi_0
entrypoints=0.3=pypi_0
h5py=2.10.0=pypi_0
idna=3.3=pypi_0
imageio=2.9.0=pypi_0
imgviz=1.4.0=pypi_0
importlib-metadata=4.8.1=pypi_0
intel-openmp=2021.3.0=h06a4308_3350
ipykernel=6.4.1=pypi_0
ipython=7.27.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.5=pypi_0
jedi=0.18.0=pypi_0
jinja2=3.0.1=pypi_0
joblib=1.1.0=pypi_0
json5=0.9.6=pypi_0
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.0.3=pypi_0
jupyter-console=6.4.0=pypi_0
jupyter-core=4.8.1=pypi_0
jupyter-packaging=0.11.0=pypi_0
jupyter-server=1.11.1=pypi_0
jupyterlab=3.2.1=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-server=2.8.2=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
kiwisolver=1.1.0=pypi_0
labelme=4.5.13=pypi_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libstdcxx-ng=9.3.0=hd4cf53a_17
libuv=1.40.0=h7b6447c_0
markupsafe=2.0.1=pypi_0
matplotlib=3.2.2=pypi_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.3.0=h06a4308_520
nbclassic=0.3.4=pypi_0
nbclient=0.5.4=pypi_0
nbconvert=6.1.0=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pypi_0
ninja=1.10.2=hff7bd54_1
notebook=6.4.4=pypi_0
numpy=1.19.2=pypi_0
open3d=0.13.0=pypi_0
opencv-python=4.5.4.58=pypi_0
openssl=1.1.1l=h7f8727e_0
packaging=21.0=pypi_0
pandas=1.3.4=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.2.0=pypi_0
pip=21.1.3=py37h06a4308_0
plyfile=0.7.1=pypi_0
prometheus-client=0.11.0=pypi_0
prompt-toolkit=3.0.20=pypi_0
ptyprocess=0.7.0=pypi_0
pycparser=2.20=pypi_0
pygments=2.10.0=pypi_0
pyparsing=2.4.6=pypi_0
pyqt5=5.15.2=pypi_0
pyqt5-sip=12.9.0=pypi_0
pyrsistent=0.18.0=pypi_0
python=3.7.7=hcff3b4d_5
python-dateutil=2.8.1=pypi_0
pytz=2021.3=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=22.3.0=pypi_0
qtconsole=5.1.1=pypi_0
qtpy=1.11.1=pypi_0
readline=8.1=h27cfd23_0
requests=2.26.0=pypi_0
requests-unixsocket=0.2.0=pypi_0
scikit-learn=1.0.1=pypi_0
scipy=1.5.3=pypi_0
send2trash=1.8.0=pypi_0
setuptools=52.0.0=py37h06a4308_0
six=1.14.0=pypi_0
sniffio=1.2.0=pypi_0
sqlite=3.36.0=hc218d9a_0
termcolor=1.1.0=pypi_0
terminado=0.12.1=pypi_0
testpath=0.5.0=pypi_0
threadpoolctl=3.0.0=pypi_0
tk=8.6.10=hbc83047_0
tomlkit=0.7.2=pypi_0
torch=1.8.0+cu111=pypi_0
torchvision=0.9.0+cu111=pypi_0
tornado=6.1=pypi_0
tqdm=4.50.2=pypi_0
traitlets=5.1.0=pypi_0
typing-extensions=3.10.0.0=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.7=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.2.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
xz=5.2.5=h7b6447c_0
zipp=3.5.0=pypi_0
zlib=1.2.11=h7b6447c_3
